+++ ChatForYou/Configurator/NavigationConfigurator.swift
@@ -1,0 +1,22 @@
+//
+//  NavigationConfigurator.swift
+//  ChatForYou
+//
+//  Created by admin on 14.06.2022.
+//
+
+import SwiftUI
+
+struct NavigationConfigurator: UIViewControllerRepresentable {
+    var configure: (UINavigationController) -> Void = { _ in }
+
+    func makeUIViewController(context: UIViewControllerRepresentableContext<NavigationConfigurator>) -> UIViewController {
+        UIViewController()
+    }
+    func updateUIViewController(_ uiViewController: UIViewController, context: UIViewControllerRepresentableContext<NavigationConfigurator>) {
+        if let nc = uiViewController.navigationController {
+            self.configure(nc)
+        }
+    }
+
+}
